Voting now Partially works
Edited the Vote Controller and took a lot of inspiration from the Report Repository.

When you have a RelatedReport as the forign key. You need to assign it to the related report and not leave it null. So you assign the forign key id and the relatedreport for it to be full circle.